requirements.txt
Flask>=2.3.0,<2.4.0
python-dotenv==1.0.0
gunicorn==21.2.0
sendgrid==6.11.0
